pyproject.toml
[tool.poetry]
name = "facial-expression"
version = "0.1.0"
description = ""
authors = ["Abhijeet Sasmal"]

[tool.poetry.dependencies]
python = "3.10.13"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
tensorflow = "^2.14.0"
black = "^23.11.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.mccabe]
max-complexity = 10
